+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/ChinookAppPanel.java
@@ -180,7 +180,7 @@ public final class ChinookAppPanel extends EntityApplicationPanel<ChinookAppMode
     EntityPanel.TOOLBAR_CONTROLS.set(true);
     EntityPanel.USE_FRAME_PANEL_DISPLAY.set(true);
     // Add a CTRL modifier to the DELETE key shortcut for table panels
-    EntityTablePanel.KeyboardShortcuts.DELETE_SELECTED.keyStroke()
+    EntityTablePanel.KEYBOARD_SHORTCUTS.keyStroke(EntityTablePanel.KeyboardShortcut.DELETE_SELECTED)
             .map(keyStroke -> getKeyStroke(keyStroke.getKeyCode(), CTRL_DOWN_MASK));
     FilteredTable.AUTO_RESIZE_MODE.set(JTable.AUTO_RESIZE_ALL_COLUMNS);
     EntityTablePanel.COLUMN_SELECTION.set(EntityTablePanel.ColumnSelection.MENU);