+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/builder/ComponentBuilder.java
@@ -24,11 +24,11 @@ import is.codion.common.state.StateObserver;
 import is.codion.common.value.Value;
 import is.codion.common.value.Value.Validator;
 import is.codion.common.value.ValueObserver;
-import is.codion.swing.common.ui.KeyEvents;
 import is.codion.swing.common.ui.component.scrollpane.ScrollPaneBuilder;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.common.ui.control.Control;
 import is.codion.swing.common.ui.control.Controls;
+import is.codion.swing.common.ui.key.KeyEvents;
 
 import javax.swing.JComponent;
 import javax.swing.JLabel;