+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/DefaultTextFieldBuilder.java
@@ -19,9 +19,9 @@
 package is.codion.swing.common.ui.component.text;
 
 import is.codion.common.value.Value;
-import is.codion.swing.common.ui.KeyEvents;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.common.ui.dialog.SelectionDialogBuilder.Selector;
+import is.codion.swing.common.ui.key.KeyEvents;
 
 import javax.swing.AbstractAction;
 import javax.swing.Action;