+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DialogBuilder.java
@@ -19,7 +19,7 @@
 package is.codion.swing.common.ui.dialog;
 
 import is.codion.common.value.ValueObserver;
-import is.codion.swing.common.ui.KeyEvents;
+import is.codion.swing.common.ui.key.KeyEvents;
 
 import javax.swing.ImageIcon;
 import java.awt.Component;