+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/key/KeyEventsTest.java
@@ -16,7 +16,7 @@
  *
  * Copyright (c) 2020 - 2023, Björn Darri Sigurðsson.
  */
-package is.codion.swing.common.ui;
+package is.codion.swing.common.ui.key;
 
 import is.codion.swing.common.ui.control.Control;