+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityControls.java
@@ -26,11 +26,11 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.exception.ValidationException;
-import is.codion.swing.common.ui.KeyEvents;
 import is.codion.swing.common.ui.Utilities;
 import is.codion.swing.common.ui.component.Components;
 import is.codion.swing.common.ui.control.Control;
 import is.codion.swing.common.ui.dialog.Dialogs;
+import is.codion.swing.common.ui.key.KeyEvents;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 import is.codion.swing.framework.model.component.EntityComboBoxModel;
 import is.codion.swing.framework.ui.EntityEditPanel;
@@ -74,7 +74,7 @@ final class EntityControls {
    */
   static Control createAddControl(EntityComboBox comboBox, Supplier<EntityEditPanel> editPanelSupplier) {
     return createAddControl(new AddEntityCommand(requireNonNull(comboBox), requireNonNull(editPanelSupplier)), comboBox,
-            EntityComboBox.KeyboardShortcuts.ADD.keyStroke().get());
+            EntityComboBox.KEYBOARD_SHORTCUTS.keyStroke(EntityComboBox.KeyboardShortcut.ADD).get());
   }
 
   /**
@@ -87,7 +87,7 @@ final class EntityControls {
    */
   static Control createAddControl(EntitySearchField searchField, Supplier<EntityEditPanel> editPanelSupplier) {
     return createAddControl(new AddEntityCommand(requireNonNull(searchField), requireNonNull(editPanelSupplier)), searchField,
-            EntitySearchField.KeyboardShortcuts.ADD.keyStroke().get());
+            EntitySearchField.KEYBOARD_SHORTCUTS.keyStroke(EntitySearchField.KeyboardShortcut.ADD).get());
   }
 
   /**
@@ -100,7 +100,8 @@ final class EntityControls {
    */
   static Control createEditControl(EntityComboBox comboBox, Supplier<EntityEditPanel> editPanelSupplier) {
     return createEditControl(new EditEntityCommand(requireNonNull(comboBox), requireNonNull(editPanelSupplier)),
-            comboBox, comboBox.getModel().selectionEmpty().not(), EntityComboBox.KeyboardShortcuts.EDIT.keyStroke().get());
+            comboBox, comboBox.getModel().selectionEmpty().not(),
+            EntityComboBox.KEYBOARD_SHORTCUTS.keyStroke(EntityComboBox.KeyboardShortcut.EDIT).get());
   }
 
   /**
@@ -113,7 +114,8 @@ final class EntityControls {
    */
   static Control createEditControl(EntitySearchField searchField, Supplier<EntityEditPanel> editPanelSupplier) {
     return createEditControl(new EditEntityCommand(requireNonNull(searchField), requireNonNull(editPanelSupplier)),
-            searchField, searchField.model().selectionEmpty().not(), EntitySearchField.KeyboardShortcuts.EDIT.keyStroke().get());
+            searchField, searchField.model().selectionEmpty().not(),
+            EntitySearchField.KEYBOARD_SHORTCUTS.keyStroke(EntitySearchField.KeyboardShortcut.EDIT).get());
   }
 
   static String validateButtonLocation(String buttonLocation) {